+++ packages/mcp/CHANGELOG.md
@@ -1,3 +1,24 @@
+## [0.23.4](https://github.com/salesforcecli/mcp/compare/0.23.3...0.23.4) (2025-10-06)
+
+
+
+## [0.23.3](https://github.com/salesforcecli/mcp/compare/0.23.2...0.23.3) (2025-10-03)
+
+
+### Bug Fixes
+
+* temp tool rename fix ([#278](https://github.com/salesforcecli/mcp/issues/278)) ([3bcaa5d](https://github.com/salesforcecli/mcp/commit/3bcaa5d4d9edbe5395e40f8550ea7effe098b18c))
+
+
+
+## [0.23.2](https://github.com/salesforcecli/mcp/compare/0.23.1...0.23.2) (2025-10-03)
+
+
+
+## [0.23.1](https://github.com/salesforcecli/mcp/compare/0.23.0...0.23.1) (2025-10-03)
+
+
+
 # [0.23.0](https://github.com/salesforcecli/mcp/compare/0.22.1...0.23.0) (2025-10-01)